+++ src/examples/zkapps/token_with_proofs.ts
@@ -62,7 +62,7 @@ class TokenContract extends SmartContract {
     receiverAddress: PublicKey,
     callback: Experimental.Callback<any>
   ) {
-    let senderAccountUpdate = this.experimental.authorize(
+    let senderAccountUpdate = this.experimental.approve(
       callback,
       AccountUpdate.Layout.AnyChildren
     );
@@ -84,14 +84,14 @@ class TokenContract extends SmartContract {
 }
 
 class ZkAppB extends SmartContract {
-  @method authorizeSend() {
+  @method approveSend() {
     let amount = UInt64.from(1_000);
     this.balance.subInPlace(amount);
   }
 }
 
 class ZkAppC extends SmartContract {
-  @method authorizeSend() {
+  @method approveSend() {
     let amount = UInt64.from(1_000);
     this.balance.subInPlace(amount);
   }
@@ -168,17 +168,17 @@ tx = await Local.transaction(feePayer, () => {
 await tx.prove();
 await tx.send();
 
-console.log('authorize send from zkAppB');
+console.log('approve send from zkAppB');
 tx = await Local.transaction(feePayer, () => {
-  let authorizeSendingCallback = Experimental.Callback.create(
+  let approveSendingCallback = Experimental.Callback.create(
     zkAppB,
-    'authorizeSend',
+    'approveSend',
     []
   );
   // we call the token contract with the callback
-  tokenZkApp.sendTokens(zkAppBAddress, zkAppCAddress, authorizeSendingCallback);
+  tokenZkApp.sendTokens(zkAppBAddress, zkAppCAddress, approveSendingCallback);
 });
-console.log('authorize send (proof)');
+console.log('approve send (proof)');
 await tx.prove();
 console.log('send (proof)');
 await tx.send();
@@ -188,19 +188,19 @@ console.log(
   Mina.getBalance(zkAppCAddress, tokenId).value.toBigInt()
 );
 
-console.log('authorize send from zkAppC');
+console.log('approve send from zkAppC');
 tx = await Local.transaction(feePayer, () => {
   // Pay for tokenAccount1's account creation
   AccountUpdate.fundNewAccount(feePayer);
-  let authorizeSendingCallback = Experimental.Callback.create(
+  let approveSendingCallback = Experimental.Callback.create(
     zkAppC,
-    'authorizeSend',
+    'approveSend',
     []
   );
   // we call the token contract with the callback
-  tokenZkApp.sendTokens(zkAppCAddress, tokenAccount1, authorizeSendingCallback);
+  tokenZkApp.sendTokens(zkAppCAddress, tokenAccount1, approveSendingCallback);
 });
-console.log('authorize send (proof)');
+console.log('approve send (proof)');
 await tx.prove();
 console.log('send (proof)');
 await tx.send();