+++ src/lib/mina.ts
@@ -1,6 +1,6 @@
 // This is for an account where any of a list of public keys can update the state
 
-import { Circuit, JSONValue, Ledger, LedgerAccount } from '../snarky.js';
+import { Circuit, Ledger, LedgerAccount } from '../snarky.js';
 import { Field, Bool } from './core.js';
 import { UInt32, UInt64 } from './int.js';
 import { PrivateKey, PublicKey } from './signature.js';
@@ -57,7 +57,7 @@ interface TransactionId {
 interface Transaction {
   transaction: ZkappCommand;
   toJSON(): string;
-  toPretty(): JSONValue;
+  toPretty(): any;
   toGraphqlQuery(): string;
   sign(additionalKeys?: PrivateKey[]): Transaction;
   prove(): Promise<(Proof<ZkappPublicInput> | undefined)[]>;
@@ -957,7 +957,7 @@ async function verifyAccountUpdate(
     checkPermission(p, 'incrementNonce');
   }
 
-  // this checks for an edge case where an account update can be authorized using proofs but
+  // this checks for an edge case where an account update can be approved using proofs but
   // a) the proof is invalid (bad verification key)
   // and b) there are no state changes initiate so no permissions will be checked
   // however, if the verification key changes, the proof should still be invalid