+++ src/lib/zkapp.ts
@@ -5,7 +5,6 @@ import {
   Ledger,
   Pickles,
   Poseidon as Poseidon_,
-  JSONValue,
   Provable,
 } from '../snarky.js';
 import {
@@ -759,28 +758,28 @@ class SmartContract {
         return zkapp.self.token();
       },
       /**
-       * Authorize an account update or callback. This will include the account update in the zkApp's public input,
+       * Approve an account update or callback. This will include the account update in the zkApp's public input,
        * which means it allows you to read and use its content in a proof, make assertions about it, and modify it.
        *
        * If this is called with a callback as the first parameter, it will first extract the account update produced by that callback.
        * The extracted account update is returned.
        *
        * ```ts
-       * \@method myAuthorizingMethod(callback: Callback) {
-       *   let authorizedUpdate = this.experimental.authorize(callback);
+       * \@method myApprovingMethod(callback: Callback) {
+       *   let approvedUpdate = this.experimental.approve(callback);
        * }
        * ```
        *
-       * Under the hood, "authorizing" just means that the account update is made a child of the zkApp in the
+       * Under the hood, "approving" just means that the account update is made a child of the zkApp in the
        * tree of account updates that forms the transaction.
-       * The second parameter `layout` allows you to also make assertions about the authorized update's _own_ children,
+       * The second parameter `layout` allows you to also make assertions about the approved update's _own_ children,
        * by specifying a certain expected layout of children. See {@link AccountUpdate.Layout}.
        *
        * @param updateOrCallback
        * @param layout
-       * @returns The account update that was authorized (needed when passing in a Callback)
+       * @returns The account update that was approved (needed when passing in a Callback)
        */
-      authorize(
+      approve(
         updateOrCallback: AccountUpdate | Callback<any>,
         layout?: AccountUpdatesLayout
       ) {
@@ -791,7 +790,7 @@ class SmartContract {
                 AccountUpdate,
                 () => updateOrCallback.accountUpdate
               );
-        zkapp.self.authorize(accountUpdate, layout);
+        zkapp.self.approve(accountUpdate, layout);
         return accountUpdate;
       },
     };